# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Power Vital API",
    "version": "1.0.0",
    "description": "Power Vital e-commerce + MLM platform backend.\n\n## Languages\nAll user-facing strings live in three locales: **TR** (primary), **RU**, **KG**.\nThe active locale is negotiated via the `Accept-Language` header and a `lang`\nquery param. AI-generated translations fill missing locale fields automatically.\n\n## Auth\nMutating endpoints require a `Bearer` JWT in the `Authorization` header.\nAdmin-only routes additionally check the `role` claim.\n\n## Rate limiting\nAll `/api/*` routes are rate-limited to 150 req/min per IP. Burst over\nthe limit returns 429 with `Retry-After`.\n\n## Errors\nAll errors return `{ error: string, issues?: ValidationIssue[] }` where\n`ValidationIssue` has the shape `{ path, message, code }`."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://powervital.kg",
      "description": "Production (Kyrgyzstan)"
    },
    {
      "url": "https://powervital.org",
      "description": "Production (intl alias)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from POST /api/v1/auth/login"
      }
    },
    "schemas": {
...
      "TrendsRange": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "days": {
            "type": "integer"
          }
        },
        "required": [
          "from",
          "to",
          "days"
        ]
      },
      "CategoryAnalytics": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "iconEmoji": {
            "type": [
              "string",
              "null"
            ]
          },
          "productCount": {
            "type": "integer"
          },
          "unitsSold": {
            "type": "integer"
          },
          "revenueKgs": {
            "type": "number"
          },
          "sharePct": {
            "type": "number",
            "description": "0-100, with 1 decimal"
          }
        },
        "required": [
          "id",
          "name",
          "productCount",
          "unitsSold",
          "revenueKgs",
          "sharePct"
        ]
      },
      "TopCustomer": {
        "type": "object",
        "properties": {
          "rank": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "orderCount": {
            "type": "integer"
          },
          "totalKgs": {
            "type": "number"
          }
        },
        "required": [
          "rank",
          "id",
          "name",
          "email",
          "role",
          "orderCount",
          "totalKgs"
        ]
      },
      "TopProduct": {
        "type": "object",
        "properties": {
          "rank": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "barcode": {
            "type": "string"
          },
          "categoryName": {
            "type": [
              "string",
              "null"
            ]
          },
          "unitsSold": {
            "type": "integer"
          },
          "revenueKgs": {
            "type": "number"
          }
        },
        "required": [
          "rank",
          "id",
          "name",
          "barcode",
          "categoryName",
          "unitsSold",
          "revenueKgs"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "List active categories (public)",
        "responses": {
          "200": {
            "description": "Categories"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "Create a category (admin)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "pattern": "^[a-z0-9-]+$"
                  },
                  "iconEmoji": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 20
                  },
                  "imageUrl": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uri"
                  },
                  "sortOrder": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1000,
                    "default": 0
                  },
                  "translations": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "object",
                        "additionalProperties": {}
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "required": [
                  "name",
                  "slug"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Validation failed"
          }
        }
      }
    },
...
    "/api/v1/admin/analytics/categories": {
      "get": {
        "tags": [
          "Admin — Analytics"
        ],
        "description": "Revenue + units sold grouped by product category over the last N days. Each entry includes sharePct (out of 100). Cancelled orders are excluded.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 365,
              "default": 30
            },
            "required": false,
            "name": "days",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Per-category revenue + sharePct + totalRevenueKgs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "range": {
                      "$ref": "#/components/schemas/TrendsRange"
                    },
                    "categories": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CategoryAnalytics"
                      }
                    },
                    "totalRevenueKgs": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "range",
                    "categories",
                    "totalRevenueKgs"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          }
        }
      }
    },
    "/api/v1/admin/analytics/top-customers": {
      "get": {
        "tags": [
          "Admin — Analytics"
        ],
        "description": "Top N customers ranked by total spend (KGS). Default limit 10, max 50.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 365,
              "default": 30
            },
            "required": false,
            "name": "days",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked customers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "range": {
                      "$ref": "#/components/schemas/TrendsRange"
                    },
                    "customers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TopCustomer"
                      }
                    }
                  },
                  "required": [
                    "range",
                    "customers"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          }
        }
      }
    },
    "/api/v1/admin/analytics/top-products": {
      "get": {
        "tags": [
          "Admin — Analytics"
        ],
        "description": "Top N products ranked by units sold (with revenue as secondary metric).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 365,
              "default": 30
            },
            "required": false,
            "name": "days",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50,
              "default": 10
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Ranked products",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "range": {
                      "$ref": "#/components/schemas/TrendsRange"
                    },
                    "products": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TopProduct"
                      }
                    }
                  },
                  "required": [
                    "range",
                    "products"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          }
        }
      }
    },
...
  },
...
}